Set a default value for PYTHON, and unset PYTHONHOME by default
Fixes Windows clients and loaders build, which uses the default system python.

PYTHONHOME was being set to the empty string when BLD_ARCH was win32 (or
another arch without an override for PYTHONHOME).
Setting PYTHONHOME to the empty string causes python not to find its
home, resulting in an error like "distutils module not found".

Signed-off-by: Jim Doty <jdoty@pivotal.io>